rename channel bank->store and change its interface
